requirements.txt
streamlit
pymupdf
google-genai
python-dotenv
fpdf2
